Fixed dirs not being created before model installs

+++ rvc.py
@@ -5,6 +5,7 @@ import torch
 import librosa
 from fairseq import checkpoint_utils
 from scipy.io import wavfile
+import numpy as np
 
 from infer_pack.models import (
     SynthesizerTrnMs256NSFsid,
@@ -129,8 +130,9 @@ def get_vc(device, is_half, config, model_path):
     vc = VC(tgt_sr, config)
     return cpt, version, net_g, tgt_sr, vc
 
+
 def rvc_infer(index_path, index_rate, input_path, output_path, pitch_change, f0_method, cpt, version, net_g, filter_radius, tgt_sr, rms_mix_rate, protect, crepe_hop_length, vc, hubert_model):
-    audio, sr = librosa.load(input_path, sr=16000)
+    audio, sr = librosa.load(input_path, sr=16000, dtype=np.float32,)
     times = [0, 0, 0]
     if_f0 = cpt.get('f0', 1)
